dockerExample/requirements.txt
Flask
opentelemetry-sdk 
opentelemetry-api
opentelemetry-exporter-otlp-proto-http 
